mid refactor

+++ src/sqlite/page/index_interior.rs
@@ -5,7 +5,7 @@ use super::page_header::PageHeader;
 pub struct IndexInteriorPage {
 
 
-    pub page_number: i64,
+    pub page_number: u32,
     pub header:PageHeader,
     pub right_cell: u32,
     pub cell_pointers:Vec<u16>
